requirements.txt
# Data processing
pandas>=1.5.0
numpy>=1.21.0

# API/Networking
requests>=2.28.0
google-api-python-client>=2.80.0

# Database
sqlalchemy>=1.4.0
# psycopg2-binary>=2.9.0  # For PostgreSQL

# Orchestration (if using Airflow)
apache-airflow>=2.5.0

# Analysis
matplotlib>=3.5.1
seaborn>=0.13.2

# Testing
# pytest>=7.0.0
# pytest-mock>=3.0.0
